WelcomePage: always update Locale
Issue #3 keeps on giving

There is an edge case where the device is clean flashed with ROM and GApps and
somehow boots directly into a different-than-English language (in our case, greek),
even without a SIM installed.
I suspect one of the GApps (probably Play Services or GoogleSetupWizard) has a subroutine
where it changes the Locale based on GPS information.

This patch forces the check so even in this case, the English locale is added to the list
of Locales.